fastapi_app/requirements.txt
fastapi
uvicorn
psycopg[binary]
yfinance
python-dotenv